Extract OrderPartiallyRefundedStateResolver
