+++ src/Resources/config/services.xml
@@ -59,6 +59,12 @@
             <argument type="service" id="sylius.repository.order" />
         </service>
 
+        <service id="Sylius\RefundPlugin\StateResolver\OrderPartiallyRefundedStateResolver">
+            <argument type="service" id="sylius.repository.order" />
+            <argument type="service" id="sm.factory" />
+            <argument type="service" id="sylius.manager.order" />
+        </service>
+
         <service id="Sylius\RefundPlugin\Checker\OrderFullyRefundedTotalChecker">
             <argument type="service" id="Sylius\RefundPlugin\Provider\OrderRefundedTotalProvider" />
         </service>
