+++ src/Resources/config/services/event_bus.xml
@@ -12,9 +12,7 @@
         </service>
 
         <service id="Sylius\RefundPlugin\Listener\UnitRefundedEventListener">
-            <argument type="service" id="sylius.repository.order" />
-            <argument type="service" id="sm.factory" />
-            <argument type="service" id="sylius.manager.order" />
+            <argument type="service" id="Sylius\RefundPlugin\StateResolver\OrderPartiallyRefundedStateResolver" />
             <tag name="messenger.message_handler" />
         </service>
